+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/omni/flux/stage_manager/plugin/interaction/usd/all_prims.py
@@ -32,6 +32,7 @@ class AllPrimsInteractionPlugin(_StageManagerUSDInteractionPlugin):
         "IsCaptureFilterPlugin",
         "OmniPrimsFilterPlugin",
         "SearchFilterPlugin",
+        "IsCategoryFilterPlugin",
     ]
     # TODO StageManager: We have LSS plugin names in the flux ext because of this system
     compatible_widgets: list[str] = [
@@ -40,4 +41,5 @@ class AllPrimsInteractionPlugin(_StageManagerUSDInteractionPlugin):
         "IsCaptureStateWidgetPlugin",
         "IsVisibleActionWidgetPlugin",
         "PrimTreeWidgetPlugin",
+        "IsCategoryHiddenStateWidgetPlugin",
     ]